h2load: Refactor rate mode
This change implements plan B described in GH-382 to simplify rate
mode, but retain its functionality as much as possible.

In this change, we removed -C option.  Instead, -c option is used to
specify the number of connections to be made, and it is now required
argument if more than 1 clients are required (this is usually the
case).  The number of requests made per connection is calculated
simply by -n / -c.